Refactor: DeliveryType 수정하였습니다.
