--- src/main/java/com/ssgsakk/ssgdotcom/common/utils/DeliveryType.java
@@ -1,5 +1,0 @@
-package com.ssgsakk.ssgdotcom.common.utils;
-
-public enum DeliveryType {
-    일반배송, 쓱배송, 새벽배송, 트레이더스쓱배송, 백화점
-}
